# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Farmers Module Service with Workflow-Based Architecture for Farm Management",
        "title": "Farmers Module API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/kisansathi/assign": {
            "post": {
                "description": "Assign a KisanSathi user to a specific farmer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kisansathi"
                ],
                "summary": "Assign KisanSathi to farmer",
                "parameters": [
                    {
                        "description": "KisanSathi assignment data",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/kisansathi/assignment/{farmer_id}": {
            "get": {
                "description": "Retrieve the KisanSathi assignment for a specific farmer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kisansathi"
                ],
                "summary": "Get KisanSathi assignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Farmer ID",
                        "name": "farmer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/kisansathi/reassign": {
            "post": {
                "description": "Reassign a KisanSathi to a different farmer or remove the assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kisansathi"
                ],
                "summary": "Reassign or remove KisanSathi",
                "parameters": [
                    {
                        "description": "KisanSathi reassignment data",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}